day1-finished

+++ codeforces/ladder-2/xenia-and-ringroad.cpp
@@ -4,9 +4,9 @@
 using namespace std;
 int main()
 {
-	int n,m;
-	int time_taken = 0 ;
-	int current_house = 1;
+	long n,m;
+	long long time_taken = 0 ;
+	long current_house = 1;
 	cin>>n>>m;
 	int *arr = new int[m];
 	for (int i = 0; i < m; i++)
